+++ plugins/flux/src/components/panel/PanelList.tsx
@@ -1,4 +1,3 @@
-import { Card, CardBody, CardTitle } from '@patternfly/react-core';
 import { IRow, Table, TableBody, TableHeader } from '@patternfly/react-table';
 import React, { useContext } from 'react';
 import { useQuery } from 'react-query';
@@ -7,21 +6,25 @@ import { ClustersContext, IClusterContext, emptyState } from '@kobsio/plugin-cor
 import Details from '../panel/details/Details';
 import { TApiType } from '../../utils/interfaces';
 
-interface IPageListProps {
+interface IPanelListProps {
   name: string;
-  cluster: string;
-  type: TApiType;
   title: string;
+  cluster: string;
+  namespace: string;
+  selector?: string;
+  type: TApiType;
   showDetails?: (details: React.ReactNode) => void;
 }
 
-const PageList: React.FunctionComponent<IPageListProps> = ({
+const PanelList: React.FunctionComponent<IPanelListProps> = ({
   name,
-  cluster,
-  type,
   title,
+  type,
+  cluster,
+  namespace,
+  selector,
   showDetails,
-}: IPageListProps) => {
+}: IPanelListProps) => {
   const clustersContext = useContext<IClusterContext>(ClustersContext);
   const resource =
     clustersContext.resources && clustersContext.resources.hasOwnProperty(type)
@@ -29,7 +32,7 @@ const PageList: React.FunctionComponent<IPageListProps> = ({
       : undefined;
 
   const { isError, isLoading, error, data, refetch } = useQuery<IRow[], Error>(
-    ['flux/list', name, cluster, type],
+    ['flux/list', name, cluster, type, namespace, selector],
     async () => {
       try {
         if (!resource) {
@@ -37,7 +40,9 @@ const PageList: React.FunctionComponent<IPageListProps> = ({
         }
 
         const response = await fetch(
-          `/api/plugins/resources/resources?cluster=${cluster}&resource=${resource.resource}&path=/apis/${resource.path}`,
+          `/api/plugins/resources/resources?cluster=${cluster}&namespace=${namespace}${
+            selector ? `&paramName=labelSelector&param=${selector}` : ''
+          }&resource=${resource.resource}&path=/apis/${resource.path}`,
           { method: 'get' },
         );
         const json = await response.json();
@@ -66,43 +71,38 @@ const PageList: React.FunctionComponent<IPageListProps> = ({
   };
 
   return (
-    <Card isCompact={true}>
-      <CardTitle>{title}</CardTitle>
-      <CardBody>
-        <Table
-          aria-label={title}
-          variant="compact"
-          borders={false}
-          isStickyHeader={true}
-          cells={resource?.columns || ['']}
-          rows={
-            data && data.length > 0 && data[0].cells?.length === resource?.columns.length
-              ? data
-              : emptyState(resource?.columns.length || 3, isLoading, isError, error)
-          }
-        >
-          <TableHeader />
-          <TableBody
-            onRowClick={
-              showDetails && resource && data && data.length > 0 && data[0].cells?.length === resource.columns.length
-                ? (e, row, props, data): void =>
-                    showDetails(
-                      <Details
-                        name={name}
-                        type={type}
-                        request={resource}
-                        resource={row}
-                        close={(): void => showDetails(undefined)}
-                        refetch={refetchhWithDelay}
-                      />,
-                    )
-                : undefined
-            }
-          />
-        </Table>
-      </CardBody>
-    </Card>
+    <Table
+      aria-label={title}
+      variant="compact"
+      borders={false}
+      isStickyHeader={true}
+      cells={resource?.columns || ['']}
+      rows={
+        data && data.length > 0 && data[0].cells?.length === resource?.columns.length
+          ? data
+          : emptyState(resource?.columns.length || 3, isLoading, isError, error)
+      }
+    >
+      <TableHeader />
+      <TableBody
+        onRowClick={
+          showDetails && resource && data && data.length > 0 && data[0].cells?.length === resource.columns.length
+            ? (e, row, props, data): void =>
+                showDetails(
+                  <Details
+                    name={name}
+                    type={type}
+                    request={resource}
+                    resource={row}
+                    close={(): void => showDetails(undefined)}
+                    refetch={refetchhWithDelay}
+                  />,
+                )
+            : undefined
+        }
+      />
+    </Table>
   );
 };
 
-export default PageList;
+export default PanelList;